test(server): relocate backend _IT tests (GameStateLastMoveResponse, LeaveGame) to ApiTests
